fivetran-destination: Fix issues we found in dogfooding (#25102)
We set up our Fivetran destination for the first time and found a few
small issues, which this PR addresses.

1. The permissions check did not consider superuser status.
2. `TruncateRequest`s can occur even if the table hasn't been created
yet, so we need to no-op in that scenario.
3. The `_fivetran_deleted` column is always required, but not always
provided, so we make sure to include it when missing.

Also if we ever get an `AlterTableRequest` we make sure to include the
debug representation in the error message, so it's easier to debug.

Note: I tried to add tests for these scenarios but it's not possible at
the moment. The tester seemed to ignore truncate requests for tables
that didn't exist and the tester always provides the `_fivetran_deleted`
columns I couldn't exercise either of these situations.

### Motivation

Fixes issues we found when debugging:
https://materializeinc.slack.com/archives/C043QH7C0VB/p1707427816936459

### Checklist

- [ ] This PR has adequate test coverage / QA involvement has been duly
considered.
- [ ] This PR has an associated up-to-date [design
doc](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/README.md),
is a design doc
([template](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/00000000_template.md)),
or is sufficiently small to not require a design.
  <!-- Reference the design in the description. -->
- [ ] If this PR evolves [an existing `$T ⇔ Proto$T`
mapping](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/command-and-response-binary-encoding.md)
(possibly in a backwards-incompatible way), then it is tagged with a
`T-proto` label.
- [ ] If this PR will require changes to cloud orchestration or tests,
there is a companion cloud PR to account for those changes that is
tagged with the release-blocker label
([example](https://github.com/MaterializeInc/cloud/pull/5021)).
<!-- Ask in #team-cloud on Slack if you need help preparing the cloud
PR. -->
- [x] This PR includes the following [user-facing behavior
changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note):
  - Fixes issues in the Fivetran destination.

+++ src/fivetran-destination/src/destination/dml.rs
@@ -50,6 +50,29 @@ pub async fn handle_truncate_table(request: TruncateRequest) -> Result<(), OpErr
         SystemTime::UNIX_EPOCH + Duration::new(secs, nanos)
     };
 
+    let (_dbname, client) = config::connect(request.configuration).await?;
+
+    let exists_stmt = format!(
+        r#"
+        SELECT EXISTS(
+            SELECT 1 FROM mz_tables t
+            LEFT JOIN mz_schemas s
+            ON t.schema_id = s.id
+            WHERE s.name = $1 AND t.name = $2
+        )"#
+    );
+    let exists: bool = client
+        .query_one(&exists_stmt, &[&request.schema_name, &request.table_name])
+        .await
+        .map(|row| row.get(0))
+        .context("checking existence")?;
+
+    // Truncates can happen at any point in time, even if the table hasn't been created yet. We
+    // want to no-op in this case.
+    if !exists {
+        return Ok(());
+    }
+
     let sql = match request.soft {
         None => format!(
             "DELETE FROM {}.{} WHERE {} < $1",
@@ -65,9 +88,11 @@ pub async fn handle_truncate_table(request: TruncateRequest) -> Result<(), OpErr
             escape::escape_identifier(&request.synced_column),
         ),
     };
+    client
+        .execute(&sql, &[&delete_before])
+        .await
+        .context("truncating")?;
 
-    let (_dbname, client) = config::connect(request.configuration).await?;
-    client.execute(&sql, &[&delete_before]).await?;
     Ok(())
 }
